move setOnDestroy to external
